Refactor BillingKtx into smaller components
The monolithic `BillingKtx` class has been split into three distinct components to improve separation of concerns:

-   `BillingConnection`: Manages the connection to the `BillingClient` and observes purchase updates.
-   `BillingRepository`: Handles data-related operations like querying purchases, product details, and acknowledging/consuming products.
-   `BillingFlowLauncher`: Manages UI-related tasks such as launching the billing flow and showing in-app messages.

The sample app has been updated to use these new components.
