+++ billingKtx/src/main/java/com/appsci/billingktx/client/connection/BillingConnection.kt
@@ -1,0 +1,12 @@
+package com.appsci.billingktx.client.connection
+
+import com.android.billingclient.api.BillingClient
+import com.appsci.billingktx.client.PurchasesUpdate
+import kotlinx.coroutines.flow.Flow
+
+interface BillingConnection {
+
+    fun connect(): Flow<BillingClient>
+
+    fun observeUpdates(): Flow<PurchasesUpdate>
+}
