+++ billingKtx/src/main/java/com/appsci/billingktx/lifecycle/BillingKtxLifecycle.kt
@@ -4,13 +4,13 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import com.appsci.billingktx.client.BillingKtx
+import com.appsci.billingktx.client.connection.BillingConnection
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
-fun BillingKtx.keepConnection(
+fun BillingConnection.keepConnection(
     lifecycleOwner: LifecycleOwner,
     state: Lifecycle.State = Lifecycle.State.STARTED,
 ) {
